# openapi-run.yaml
swagger: '2.0'
info:
  title: GLEAM SEIR Endpoint
  version: '0.1.0'
host: gleam-seir-api-883627921778.us-west1.run.app
schemes:
  - https
produces:
  - application/json
x-google-backend:
  address: https://gleam-seir-api-883627921778.us-west1.run.app
  protocol: h2
x-google-management:
  metrics:
    - name: "Num_root_reads"
      displayName: "Num_root_reads"
      valueType: INT64
      metricKind: DELTA
  quota:
    limits:
      - name: "LIMIT-Num-root-reads"
        metric: "Num_root_reads"
        unit: "1/min/{project}"
        values:
          STANDARD: 2
paths:
  /:
    get:
      summary: Test
      operationId: test__get
      responses:
        '200':
          description: Successful Response
          schema:
            type: object
      x-google-quota:
        metricCosts:
          Num_root_reads: 1
      security:
        - APIKeyHeader: []
  /original:
    post:
      summary: Original
      operationId: original_original_post
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Params'
      responses:
        '200':
          description: Successful Response
          schema:
            type: object
        '422':
          description: Validation Error
          schema:
            $ref: '#/definitions/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /multiple:
    post:
      summary: Multiple
      operationId: multiple_multiple_post
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ListParams'
      responses:
        '200':
          description: Successful Response
          schema:
            type: object
        '422':
          description: Validation Error
          schema:
            $ref: '#/definitions/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /create_compute:
    post:
      summary: Create_Compute
      operationId: create_compute_post
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Params'
      responses:
        '200':
          description: Successful Response
          schema:
            type: object
        '422':
          description: Validation Error
          schema:
            $ref: '#/definitions/HTTPValidationError'
      security:
        - APIKeyHeader: []
definitions:
  HTTPValidationError:
    type: object
    title: HTTPValidationError
    properties:
      detail:
        type: array
        items:
          $ref: '#/definitions/ValidationError'
  ListParams:
    type: object
    title: ListParams
    required:
      - days
      - sims
      - beta_epsilon
    properties:
      days:
        type: integer
        title: Days
      sims:
        type: integer
        title: Sims
      beta_epsilon:
        type: array
        items:
          type: object
        title: Beta Epsilon
  Params:
    type: object
    title: Params
    required:
      - days
      - sims
      - beta
      - epsilon
    properties:
      days:
        type: integer
        title: Days
      sims:
        type: integer
        title: Sims
      beta:
        type: number
        title: Beta
      epsilon:
        type: number
        title: Epsilon
  ValidationError:
    type: object
    title: ValidationError
    required:
      - loc
      - msg
      - type
    properties:
      loc:
        type: array
        items:
          type: string
        title: Location
      msg:
        type: string
        title: Message
      type:
        type: string
        title: Error Type
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    name: X-API-Key
    in: header